feat: add authentication and restructure project files
