+++ e2e/integration/nextjs/next.config.ts
@@ -1,7 +1,7 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+	/* config options here */
 };
 
 export default nextConfig;
