+++ e2e/integration/nextjs/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "next": "catalog:nextjs",
+    "pg": "^8.16.3",
     "react": "catalog:nextjs",
     "react-dom": "catalog:nextjs"
   },
