+++ e2e/integration/nextjs/src/app/api/auth/[...all]/route.ts
@@ -1,0 +1,4 @@
+import { auth } from "../../../../lib/auth";
+import { toNextJsHandler } from "better-auth/next-js";
+
+export const { GET, POST } = toNextJsHandler(auth);
