+++ e2e/integration/nextjs/src/app/page.tsx
@@ -1,3 +1,3 @@
 export default function Home() {
-  return <div>Hello world!</div>;
+	return <div>Hello world!</div>;
 }
