+++ e2e/integration/nextjs/src/lib/auth.ts
@@ -1,0 +1,11 @@
+import { betterAuth } from "better-auth";
+import { Pool } from "pg";
+
+export const auth = betterAuth({
+	database: new Pool({
+		connectionString: process.env.DATABASE_URL!,
+	}),
+	emailAndPassword: {
+		enabled: true,
+	},
+});
